requirements.txt
blessed>=1.20.0
google-genai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
prompt-toolkit>=3.0.0
